+++ src/components/map/mapEventHook.ts
@@ -6,7 +6,7 @@ import * as action from "../../script/actions"
 import { Line } from "../../script/line"
 import { LatLng } from "../../script/location"
 import { logger } from "../../script/logger"
-import { selectMapState } from "../../script/mapState"
+import { selectMapState } from "../../script/rootState"
 import { Station } from "../../script/station"
 import { AppDispatch } from "../../script/store"
 import { parseQueryBoolean } from "../../script/utils"
