+++ src/components/map/mapHook.ts
@@ -5,7 +5,7 @@ import * as action from "../../script/actions"
 import { Line } from "../../script/line"
 import { LatLng, MapCenter } from "../../script/location"
 import { logger } from "../../script/logger"
-import { selectMapState } from "../../script/mapState"
+import { selectMapState } from "../../script/rootState"
 import { Station } from "../../script/station"
 import { AppDispatch } from "../../script/store"
 import { PolylineProps, RectBounds, getBounds, getZoomProperty, isInsideRect } from "../../script/utils"
