+++ src/script/mapState.ts
@@ -1,15 +1,9 @@
 import { NavState } from "../components/navState"
-import { DataType } from "./StationService"
 import { CurrentLocationState, MapCenter } from "./location"
-import { Station } from "./station"
 
 export interface GlobalMapState {
   radarK: number
   showStationPin: boolean
-  /**
-   * 表示する駅データの種類
-   */
-  dataType: DataType | null
   /**
    * GPSで現在地を監視する状態
    */
@@ -26,14 +20,4 @@ export interface GlobalMapState {
    * Map中心位置
    */
   mapCenter: MapCenter
-  /**
-   * Mpa上に表示する駅一覧
-   */
-  stations: Station[]
 }
-
-export interface RootState {
-  mapState: GlobalMapState
-}
-
-export const selectMapState = (state: RootState) => state.mapState
