+++ src/script/store.ts
@@ -1,9 +1,11 @@
-import {configureStore} from "@reduxjs/toolkit"
+import { configureStore } from "@reduxjs/toolkit"
 import mapStateReducer from "./mapStateSlice"
+import stationStateReducer from "./stationStateSlice"
 
 export const store = configureStore({
   reducer: {
     mapState: mapStateReducer,
+    stationState: stationStateReducer,
   },
 })
